grepUrl: Use the tld.js library improve domain splitting.
Before this commit, grepUrl used a heuristic to split domains:
  * if it has one dot:
    cut off the bit after the dot
  * if it has two or more dots:
    cut off the bit before the first dot, and after the last dot
e.g.
  http://amazon.com/ -> amazon
  http://amazon.co.uk/ -> co
  http://www.amazon.co.uk/ -> amazon.co
  http://www.aws.amazon.co.uk/ -> aws.amazon.co

This worked pretty well most of the time.
However, as the above example shows, it can cause havoc
when trying to log on to different bits of amazon.

This commit makes grepUrl use the tld.js library to deal with domains properly:
e.g.
  http://amazon.com/ -> amazon
  http://amazon.co.uk/ -> amazon
  http://www.amazon.co.uk/ -> amazon
  http://www.aws.amazon.co.uk/ -> amazon