Use a narrower hash field on portable html page
Old way:
The hash field on the portable html page was big, and you
could see the whole hash.

Problem:
A shoulder surfer could see the hash and remember it, or take a
photo of the screen.

New way:
Use a narrower hash field, so that most of the hash is obscured.